Use Caddy 2.x syntax
CI shows that the module configure caddy with unsupported settings and
pass unsupporte command line arguments to the daemon. These where legacy
settings / options that got deprecated with Caddy 2.x and are no
removed.

Adjust the produced config so that it is valid in regard of the current
configuration specification, and fix the cli arguments to match what is
currently supported.

This fix acceptance tests.