Rework capabilities management
Setting capabilities on the binary works fine in docker but the
acceptance tests fail when running in Vagrant.  While trying to identify
the root cause of the issue, it feels inadequate to set the capabilities
on the binary itself because it allows any user on the system to run
caddy and attach it on a priviled port.

Removing the explicit file capabilities management and passing this
responsibility to systemd helps solving the above issue, and also fix
running the acceptance tests in Vagrant.