Fix installation of a specific version
The module previously never touched caddy once it had been installed.
While we legitimately do not want to update the binary when we fetch the
latest version form the internet, we want to see an acutal change when
we update the catalog configuration to change the installation method or
the specific version we want to install.