Require a version when installing from GitHub
The `caddy::version` parameter is only used when `caddy::install_method`
is "github", but no error was previously raised if the user set a
specific version that was ignored, and the resulting configuration was
not what the user expected.

In order to avoid confusion, do not default to an outdated version of
caddy when installing from GitHub (requiring the user to explicily tell
which version they want) and raise an error if a version is set but is
ignored.

This should make the module more user-friendly for new users.